bug fix  with myfirst commit
