--- README.md
@@ -1,1 +1,0 @@
-MyBrand portfolio backend part
