+++ src/controllers/auth/signup.ts
@@ -1,0 +1,23 @@
+import {Request, NextFunction, Response } from 'express';
+import  jwt  from 'jsonwebtoken';
+import { IUser } from '../../utils/userType';
+
+export const userSignup=async (req:Request, res:Response, next:NextFunction) => {
+
+  let token:string;
+  if(req.user){
+    const user=req.user as IUser
+    const body = { _id: user._id, email: user.email,names:user.names };
+    token= jwt.sign({user:body}, process.env.JWT_SECRET as string,{expiresIn: '15m'} );
+ 
+  res.status(200).json({
+    message: 'Signup successful',
+    token:`Bearer  ${token}`,
+    user:req.user
+  });
+  }else{
+    res.status(400).json({
+      maessage:"something went wrong please try again"
+    })
+  }
+}
