--- src/routers/secure-routes.ts
@@ -1,16 +1,0 @@
-import  Router, {Request, NextFunction, Response } from 'express';
-const secureRoute =Router();
-secureRoute.get(
-    '/profile',
-    (req:Request, res:Response, next:NextFunction) => {
-      console.log('here')
-      res.json({
-        message: 'You made it to the secure route',
-        user: req.user,
-        token: req.query.secret_token
-      })
-    }
-  );
-  
-  
-  export default secureRoute
