+++ src/utils/userType.ts
@@ -1,0 +1,8 @@
+import { Schema } from "mongoose";
+export interface IUser extends Document {
+    _id:Schema.Types.ObjectId,
+    names:string;
+    email:string;
+    password:string;
+    userType:string
+}
